TEMPORARY: Add expected windows test failures
This is a temporary solution to unblock other MRs. We should investigate
why these test are failing and if possible fix them.

Note that while the two test in question fail in all 4 windows
environments (1809, 2004, 20H2, 21H1), I've only added them to the 1809
list of expected failures. This is because it appears that the `check
windows xxxx integration test failures` jobs don't even run! I'll not
add the two test to the expected failures for the other three
environments so we can more easily tell when they do run (because they
should fail).