powershell: fix unwanted progress streams leaking to output
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/3707
forced all commands to be encoded in base64 to avoid UTF-8 encoding
issues.

However, this caused unwanted progress output to be leaked with
each command due to an upstream bug:
https://github.com/PowerShell/PowerShell/issues/5912.

This commit reverts he previous behavior of using unencoded commands
for `powershell`, but leaves in the encoding for `pwsh`.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/29462

Changelog: fixed