powershell: Fix stdin handling with scripts
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/3728
added logic to add braces around a script only if the file were
executed in a script, but the braces are actually needed when the
script is passed in via stdin instead of a file. As the Powershell
`-help` output says:

```
If the value of Command is a script block, the script block must
be enclosed in braces ({}). You can specify a script block only when
running PowerShell.exe in Windows PowerShell. The results of the
script block are returned to the parent shell as deserialized XML
objects, not live objects.
```

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/29551

Changelog: fixed