Fix windows integration tests failure check
The rules for when `check windows xxxx integration test failures` run
were wrong. The rules for these thee jobs were:

    default-branch-only:no_docs:always

Which means they would not run on MRs, which in turn meant that new
failures wouldn't be caught until later (i.e. release or main build).
This commit will make those jobs run on MR pipelines too.

And now because of this, the two integration tests that fail on windows
will have to be added to the expected failures.