Use JobVariable.Value() for internal values
This change introduces JobVariable.Value() that can be used for variables
that Runner reads internally, and don't need to be expanded in the same
way as a variable does when it's expanded for the job's environment.

This fixes issues where a value Runner expects to read has been set to
a file-type variable, but Runner is only interested in the contents
of the variable, and doesn't want it expanded to the pathname of where
the contents will reside within the job environment.