Log artifact download request host
When a runner attempts to download an artifact, it may be redirected
to an object storage or CDN link. The redirected link may be rejected
due to a 403 error, but from the perspective of the user it looks like
the download was outright rejected by GitLab. Some users may have
network policies that cause the redirection to fail since the host
of the redirected link may not be allowed.

To make it easier to debug these class of problems, we now annotate
the response host in the log.

Changelog: changed