[common] Expand image/container fields in CI config
It is possible and allowed for users to specify any of job image or
service image names or aliases as Variables. Currently image.name is
expanded where used, but there are many places where it and the other
fields are not expanded, which results in
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/29499

This change fixed the above issue, but only for CI config. I'll fix it
for runner configs next.