[k8s] Fix latent bug in service log capture
If the services were only specified via the job config, `s.Build.Services`
would be empty and the service log capture would not be set up. It turns out
that `s.options` is a combination of the services defined in both the job and
runner config.

I'll make sure it's tested later.