Update DOCKER_VERSION to enable proper --platform handling on import
According to https://github.com/moby/moby/pull/43103, when --platform
was initially added to docker import, it only honored the OS portion
of the argument, and Architecture was silently ignored. That PR, which
resolved the issue and made import honor the architecture segment, was
merged for 23.0.0. Therefore, in order for docker import --platform to
resolve https://gitlab.com/gitlab-org/gitlab-runner/-/issues/30869, we
must use at least Docker 23.0.0.