Change how supported failure reasons are filtered
Failure reasons are reported back to GitLab. But to be properly handled,
GitLab must know their value. Therefore tha values are part of the
API.

To properly handle that, GitLab is reporting what failure reasons are
known on its side and on the Runner side we're filtering the generated
failure reason through that list.

This allows us to introduce a new failure reason and start using it on
the Runner side, but until its added to GitLab, through the API
we send "unknown_failure".

While this mechanism was added to pass only known failure reasons to
GitLab, it is also used when putting failure reason to the metrics
exported by Runner.

There is no need for that.

This commit changes how failure reason filtering is done, so that:

- failure reasons sent to GitLab report only supported values,
- failure reason added as a label to the metric uses the correct value,
  which allow runner owners to track these as soon as they are released.

As we might in the future create more specific failure reasons that are
extracted from the scope of an exsiting one, this commit adds also
a mapping mechanism.

New reasons with a more specific scope can be mapped to longer existing
reasons with a wider scope and the first reason in that chain that is on
the list of the ones supported by GitLab will be sent.