Enable variable injection to Persistent Volume Claim name
By allowing variable injection here, we enable the user to
allow different jobs to attach to different volumes,
eg. by CI_CONCURRENT_ID. The volume claims need to be
created ahead of time by the user, but this will still
alleviate
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/27835
a bit and works as a step
towards a more comprehensive solution.