Kubernetes executor: prevent background processes from hanging the job
This commit changes the way we execute scripts when attaching to
containers in the kubernetes controller. It used to be that we executed
scripts by piping them into tee.

This works well most of the time, but if the executed script also
starts processes in the background, they'll inherit the tee pipe as
their standard output & error file descriptors. This means that when
the parent command terminates, if these background processes survive,
then the tee process never ends up getting EOF, blocking the shell we
are attaching to.

This commit changes that command so that the problematic command is now
run in a background subshell, which means we don't wait on a potentially
blocked tee.

There are better ways to do this. Normally, you'd create a session with
a controlling terminal, and on process exit the session would get
SIGHUP'd; unfortunately this is hard to do in a portable manner, and
adding new requirements to CI images seem like a no-go. What is real
however is that this issue hasn't been fixed for five years and is
impacting people, so let's just do this instead.

Co-Authored-By: Romuald Atchadé  <ratchade@gitlab.com>