Drop extension from cache filename in logs
Caches written to disk use the extension `.zip` even when a different
compression format is used. This is different to how caches are stored
in remote storage, where no extension is given.

Since we already use the magic bytes of an archive to determine the type
the extension isn't needed. However, if we dropped the extension, we'd
immediately invalidate existing caches. To avoid this, we keep the .zip
extension (since the cache store is an internal implementation detail)
but no longer log the extension simply to avoid confusion.