Make k8s API Call retryable
When executing a k8s API Call, this request can fail for multiple reasons (network issue, kube-api-server error, etc ...). In some case, the request can fail despite being executed on the server.
Among the requests made, the `POST` (Create calls) and `PUT/PATCH` (Update Calls) are the ones at risk. In order to avoid the duplication of data, the `name` of the resource to be created must be known beforehand.
Also, when the call return a `409` (Conflict) error, the resource is then retrieved and we stop the retry cycle (until the `GET` also fails)