In attach check for running processes before retrying a command
Retrying attach and exec commands is a bit different from regular Kubernetes requests.
Since the attach / exec commands are executed by opening an HTTP stream to the Kubernetes server and piping the command into that stream and then expecting a response there's no good place to check whether the whole command execution was successful.
If we check whether the Stdin stream was read - the connection might have broken up after during transit of that meaning that the command was never executed.
It could have also been broken during the reading of the response stream - meaning that it was executed, but we can't know that.
The only solution is to check for certain whether the process is already running.
For attach that is easy since the process is completely running in the background, and we receive the status of it through the log file and the log processor moves things forward.