Fix TestStackDumping flaky test and incorrect log output
When running `go test` for the log package, this appears to be a problem
of the TestStackDumping test, but is really related to TestHandleCliCtx
and how it doesn't correctly handle ending the goroutine for runtime
dumps.

In the logs, it looks like it's for TestStackDumping though due to the
timing of it. The output of the dump, running outside of when the real
test ended, messes with the test output and confuses the Splitic test
runner.