Do not install python in ubi-fips-base image
Currently when assembling the `ubi-fips-base` image, we build git from
source with the `--with-python` flag, which requires we install python,
which in turn pulls in `pip`.

https://gitlab.com/gitlab-org/gitlab-runner/-/issues/31065 reports CVE
vulnerability PRISMA-2022-0168 against `pip`. This vulnerability has not
been fixed upstream.

Meanwhile, it's not clear why we build git with `--with-python`.
Installing git via the canonical package managers in alpine, debian and
fedora does NOT pull in python, so I'm not sure why we need it here, and
that wasn't documented AFAICT.