Add RISC-V support
This adjusts the build process to support RISC-V. It includes an update
to the latest version of Gox because the stable version does not support
RISC-V.

This also adds a new alpine-edge helper image, which is required
because alpine stable does not have RISC-V support. The edge image also
uses alpine's build of git-lfs instead of the prebuilds, which will only
add RISC-V in the next release.