Only decode certificates if HTTP GET is successful
Previously when the Runner attempted to resolve the full TLS chain, it
fetched URLs from the certificate chain without paying attention to
the HTTP status code. This could cause the Runner to decode a bogus
payload if the endpoint were blocked or down.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/36319