Build git in a separate layer in ubi base image
Building gut requires a bunch of packages we neither need or want in a
final runtime image. We can keep the final image small by building git
in a separate layer, and just copying the final products over to the
final image.

This will address a number of CVE vulnerabilities.