Add test case for unsupported options
Currently `network.GitlabClient.RequestJob()` returns `(nil, -1)` on
json parse errors, and `commands.RunCommand.requestJob()` returns `(nil,
nil, nil)` before creating a job trace logger, which does not give us a
chance to report the unsupported option in either case (fail or process
job). We'll have to change all that AND this test to make this work. For
now this test captures the existing functionality.