Some base classes to facilitate executor option validation
Validating executor options will be the same for all executors and
executor options, so lets try to centralize as much of it as possible in
one type which can be extended by each concrete executor option.

The relevant things to encapsulate here are validating the options and
capturing any unsupported options in a descriptive way.