API to expose UnsupportedOptionErrors
Add UnsupportedOption() method to all relevant types in the hierarchy
from JobResponse to ImageDockerOptions and ImageDockerOptions, combining
them where appropriate.

Why do we need this? The code that requests a job helpfully discards
json parsing errors, so it's not possible to determine from callers
what, if any, parsing errors occurred. Nor do we necessarily want to
fail the job if it included unsupported options. This API lets any code
that as access to the JobResponse to inspect the result of the message
parsing, and decide what to do based on the result.

We'll use this later to log a message or fail the job when the latter
includes unsupported options.