Add missing runtime package to ubi base image
Over in
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4289 I
moved building of git in the ubi base image to a separate layer.
Building git required installing a number of packages which were
originally left behind in the final image, most of which are not
required at runtime and bloated the final image (and provided a larger
surface area for CVE vulnerabilities). Moving this step to a separate
layer had the effect of no longer installing a number of packages in the
final image, some of which it turns out are required in downstream
images at runtime. We don't know anymore which downstream images require
which packages, so to be safe let's install them in the base image.