Install git and git-lfs from package manager
The vulnerabilities against `git` and `git-lfs` that prompted us to
install them manually have been downgraded from `High` and `Critical` to
`Medium`. Since we don't address `Medium` CVE vulnerabilities, we can
return to installing these packages via microdnf.

This change not only addresses
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/36051, but also
cuts the pipeline runtime by ~3 hours when the `prepare ubi base` job is
triggered, and makes that job much less flaky since it often timed out.

This is not the end of this story though. We need to put ourselves in a
better position to handle CVEs against 3rd party packages.