Retrieve script PATH before executing script
In attach mode, we retrieve `$LASTEXITCODE` to determine if a job passes or fails. The instruction for the PATH retrieval was used instead to determine if the job has failed or not. To fix the issue, and prevent it to be introduced in the future, the script exit code is saved in a variable which is then used when sending the json.