Lock `gitlab-dangerfiles` to 4.1.0
The danger-review job runs the following line:

$ [[ ! -f "Gemfile" ]] && bundle init && bundle add gitlab-dangerfiles --version "$GITLAB_DANGERFILES_VERSION"

$GITLAB_DANGERFILES_VERSION comes from the gitlab project, and was
recently updated to `4.2.0`. For reasons yet unknown, this job fails
when run in our pipeline.

For now, pin the version of this gem to the previous know working
version to give us time to investigate why this failure happens in our
pipeline with the new version.