Integration tests for then image/service platform specified
Specify image platform for job and service image in the job, and ensure
it was used. Also ensure platform was not used for the helper image and
for the service image for which platform was not specified.

This test is a bit indirect in that it relied on the job log to assert
whether platform was used to pull the docker image, which is not the
same as directly inspecting the pulled image.

Skip this test if the docker daemon version is less than 1.41.