Fix labeling of the gitlab_runner_failed_jobs_total metric
While working on https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4098
we've refactored the way how failure reason setting is handled. When
doing that we've missed a specific case when failure reason would be
empty (some cases for `script_failure` failures and cases that are
related to job execution and for which Runner doesn't yet recognize
specific failure reason).

Previously that would be automatically translated to `script_failure`
for both the reason reported back to GitLab and the reason tracked by
metrics.

After the change this check was still executed when preparing the reason
that should be reported back to GitLab (where we need to do few more
checks) but was missed for the metric collection.

This commit will fix that.

This closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/36929