Use Init=true when creating ALL docker containers
This will run `/sbin/docker-init -- sh` as PID 1, which will correctly
propagate SIGTERM (and other signals) to child process. Note that if any
of the child processes are shells, signal propagation will effectively
stop at that process since shells have the prickly property of
swallowing signals. All this to say this running the container with a
proper init system is not enough to guarantee graceful shutdown of the
container. We'll fox that later.