Script to send SIGTERM to all processes in container
This script is a simple, portable and robust, if a bit heavy handed way
to ensure all process in a container receive SIGTERM. This script is
indented to be run in docker or kubernetes containers ONLY to ensure
graceful shutdown of build, service and potentially other containers. It
sends SIGTERM to all PIDs excluding itself and 1, in decreasing numeric
order, positing that the higher PIDs are likely the processes blocking
and thus preventing the container from shutting down cleanly. The inner
while loop waits for up to 5 seconds for the last killed PID to exit
before moving onto the next PID. Note that processes that are shells
will ignore SIGTERM anyway, so this script is not as heavy handed as it
might appear.