Use copy `dumb-init` on build container
`dumb-init` is available on all `GitLab Runner Helper` images for `Linux like system`. When creating the containers (Helper and Build), the previous `container.Command` ran cause `/bin/sh` (on build), `/bin/bash` (on helper) to be the PID 1.
This prevents the termination signals sent to PID 1 to be propagated to the child process.
In this MR, the `dumb-init` is copied in `the script dir for the attach mode` and then used to run the bash shell script at the containers creation. This allows the `dumb-init` to be `PID 1`.