Make space to pass graceful shutdown function StopwKillWait()
executor.Cleanup() is not the right place to call
sendSIGTERMToContainerProcs() as it does not handle all cases (e.g.
RUNNER_SCRIPT_TIMEOUT is exceeded) AND prevents logs made after the
timeout cancel from being included in the job trace log.

The right place to call sendSIGTERMToContainerProcs is StopwKillWait,
immediately before stopping the container. This commit makes space for
that.