Make autoscaler integration tests pass
These test started failing around the time
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4488 was
merged. The failure is in pulling the helper image, which is tagged with
the review branch's head commit, and does not exist. We theorise that
another test in the same block (x or 4) as these failing tests is
loading the helper image into the local docker daemon, and with the
tests introduced in the above MR, the test ordering/grouping was
shuffled around, and the tests that load the helper image are no longer
in the same block as these test.

FOR NOW, lets work around the lest by specifying a helper image that is
know to exist in the registry.