Fix broken main
Looks like something went wrong with the `16.6.1` release, or rather the
merge of `16-6-stable` back into `main`.

I'm not sure if this is supposed to happen but all the new (since
`16.6.0`) commits from `16-6-stable` were merged back into `main`,
including the merge commit for
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4488:
* original merge commit
  https://gitlab.com/gitlab-org/gitlab-runner/-/commit/c0cd6b7ff3e4beecff2c5e512a5a537deaa14a43
* merge commit from `16-6-stable`
  https://gitlab.com/gitlab-org/gitlab-runner/-/commit/66da9072f62594b7035969ad4ec96994cdd7077d

Note that the diff context is slightly different for each! This is
because
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4488 was
merged into main AFTER
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4325, and
the latter included as diff context from the former. But, the former was
not cherry picked into the 16-6-stable branch. So, the cherry pick
triggered a merge conflict, there resolution of which required dropping
the relevant context from
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4325. And
when
https://gitlab.com/gitlab-org/gitlab-runner/-/commit/66da9072f62594b7035969ad4ec96994cdd7077d
merged back into main, it effectively removed part of
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4325.

This commit adds back that dropped diff context to fix main. We still
should investigate why/how this happened, and if either:
* i handled the cherry pick merge conflict incorrectly, or
* there's a bug in the releaser tool.