Avoid quoting problems
The cmd shell runner was generating some absolute nonsense, using %q to
print an already-batch-escaped path. This would generate a string
looking like "\\tmp\\gitlab-runner\\build\\project.tmp\\KEY". Echo
would then complain that it couldn't find the network-mapped drive
"\\tmp", and fail.

This MR reworks the code to _first_ set the environment variable with
the path, and then uses that env var as the output file to write to.

We additionally redirect the output of echo. to NUL to avoid writing
useless empty lines to the output of the job.

The tests were adjusted. For what it's worth: The tests here are
absolutely _bonkers_. Half of them test some completely broken behavior,
while the other half have difference in behavior between windows and
linux that should really not be there. This seems to all stem from using
the go `filepath` library to handle paths, instead of specialized
routines for Windows path. This means simple stuff like checking if a
path is absolute will have a broken behavior when running on Linux.