Add AllowedUsers to Docker executor config
In https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4483 we
added the ability to configure the user with which to run a build
container, AND made that take precedence over the docker user configured
in the runner. That latter part was wrong; the runner config should take
precedence. Not only that, but as with pull_policies and
allowed_pull_policies, we need to limit the user specified in the job
config to a list of allowed users specified in the runner config.

This commit adds allowed_users to the docker configuration, and a
utility method to determine if a given user is allowed or not according
to allowed_users.

See https://gitlab.com/gitlab-org/gitlab-runner/-/issues/2750#note_1712531790