Set saner default for --config option
With the change in the previous commit, the existing default config path
no longer works. Since the default config path now depends on the
specified user, we can't set it statically anymore. We have to evaluate
it dynamically based on the specified user.

The alternative is to ONLY mention this in the documentation.