Allow zero value for cleanup_grace_period_seconds
In
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/30230#note_1638579844
a user notes that he wants to set a `cleanup_grace_period_seconds` of 0s
to trigger pod-cleanup immediately. However a value of 0 is no accepted
and the default of 10m is applied.

This change makes 0s and allowed/accepted value.

Note that this change alone does not fix https://gitlab.com/gitlab-org/gitlab-runner/-/issues/30230