Don't install wget in ubi images
The version of `wget` installed in the latest `ubi8` image is subject to
vulnerability CVE-2021-31879. The thing is, `curl` is installed by
default in `ubi8` and can do everything `wget` can, so there's really no
need for `wget`.

See https://gitlab.com/gitlab-org/gitlab-runner/-/issues/37054

Note that I have not been able to find uses of `wget` in this repo nor
in https://www.gitlab.com/gitlab-org/ci-cd/gitlab-runner-ubi-images, but
it is possible that downstream users extend these or those images and
assume/expect `wget` to be installed.