Really silence error when execing on container that does not exists
Because docker error types are in fact interfaces with a unique
identifying method, it's not possible to use errors.Is or errors.As on
them. And because we wrap those errors as they are returned up the
chain, we can't use errdefs directly. Do this instead.

See https://gitlab.com/gitlab-org/gitlab-runner/-/issues/37329#note_1793492722