Change beta to pre in version string
As discussed in
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/31098,
the `~beta` part in the version string concerns users. Especially
because it's what we print for jobs executed on GitLab.com Hosted
Runenrs.

Whatever we merge to `main` (which now will be released as
`vX.Y.Z~pre.N.gSHA`) we consider ready for the next stable release and
stable is just created on whatever is on `main` at the decision day.

We're also deploying these versions to our Hosted Runners in a form of a
rolling update, for which `pre` also makes more sense.