Update github.com/hashicorp/vault/api
What we actually want to do is update gopkg.in/square/go-jose.v2 because
it is subject to CVE-2024-28180. It is pulled in by
github.com/hashicorp/vault/api though:

    > go mod why gopkg.in/square/go-jose.v2
    gitlab.com/gitlab-org/gitlab-runner/helpers/vault
    github.com/hashicorp/vault/api
    gopkg.in/square/go-jose.v2/jwt
    gopkg.in/square/go-jose.v2

So to update go-josh we need to update hashicorp/vault.

Note that gopkg.in/square/go-jose.v2 is orphaned and has been replaced
with github.com/go-jose/go-jose/v3.