Stop installing tar in ubi fips images
There are 2 fedramp issues against the runner and helpr fips images
because of vulnerabilities in tar:

* https://gitlab.com/gitlab-org/gitlab-runner/-/issues/37263
* https://gitlab.com/gitlab-org/gitlab-runner/-/issues/37264

I haven't found anywhere this project, nor any of the downstream
projects that consume these images, a single use of tar.

In light of that I'll remove it; we can always add it back if we break
something.