Support Google Secrets from projects other than the WIF pool one
As noted in https://gitlab.com/gitlab-org/gitlab-runner/-/issues/37487,
there is currently no way to reference a secret from a project other
than the one containing the WIF pool despite Google supporting this.

Extend the interpretation of `gcp_secret_manager:name:` to allow
fully-qualified secret resource names.

If the secret name matches the pattern `projects/*/secrets/*`, we use it
as the full resource name otherwise we fall back to the existing
behaviour of interpolating the project id used to construct the WIF
audience.