Add AFTER_SCRIPT_IGNORE_ERRORS variable
By default errors in `after_script` are ignored. This is not the
desired behaviour for many uses of `after_script`, for example
if you use it to check git is clean after each step then you don't
want to ignore errors.

Arguably the behaviour should always have been to observe errors,
since ignoring errors within the script is always possible
(e.g. `./my_after_script.sh || true`) but observing them if they
are ignored by the runner is impossible.

This commit adds a variable `AFTER_SCRIPT_IGNORE_ERRORS` that can
be set to `false` to not ignore errors in all `after_script`s.
The default is `true` and it also uses `true` if there's a parse
error.